+++ glean_parser/javascript.py
@@ -11,7 +11,7 @@ Outputter to generate Javascript code for metrics.
 import enum
 import json
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Union, Callable  # noqa
+from typing import Any, Dict, Optional, Callable
 
 from . import metrics
 from . import util
@@ -123,9 +123,10 @@ def output(
         options = {}
 
     platform = options.get("platform", "webext")
-    if platform not in ["qt", "webext"]:
+    accepted_platforms = ["qt", "webext", "node"]
+    if platform not in accepted_platforms:
         raise ValueError(
-            f"Unknown platform: {platform}. Accepted platforms are qt and webext."
+            f"Unknown platform: {platform}. Accepted platforms are: {accepted_platforms}."  # noqa
         )
     version = options.get("version")
     if platform == "qt" and version is None: